Various code improvements based on CLion hints. Reworked CoordBuffer and CoordBufferImpl to CoordBuffer, because template will be instantiated now always with the same parameter after OpenGL now does not inherit from MapPainter anymore. Further optimizations are likely.

+++ libosmscout-map-opengl/src/osmscout/TextLoader.cpp
@@ -80,7 +80,7 @@ namespace osmscout {
         sumWidth += 2;
       } else {
         texture->data = new unsigned char[bit.width*bit.rows];
-        for(int t = 0; t < bit.rows*bit.width; t++)
+        for(size_t t = 0; t < bit.rows*bit.width; t++)
           texture->data[t] = bit.buffer[t];
         texture->width = bit.width;
         texture->height = bit.rows;
